When the player dies hitting a enemy all enemies are destroyed.
I've also changed the name of some variables and some functions to utilize the reference of variables, not their value.

CLion also remade the sMovement() player movement method. I barely understand how It works, but I'll look into it.

Also, new method just dropped. toFloat() has a innovative and never seem before work. It converts an unsigned int to a float. Yeah. Just because I need to call a lot of static_cast, I decided (In all truth It was Copilot) that It's better to have a centralized function that does that.

Changes on Game.cpp and Game.hpp